feat: useProductDetail 생성 및 적용 #137
