+++ src/routes/Product/Detail.tsx
@@ -1,4 +1,3 @@
-import { useEffect, useState } from "react";
 import { BsHeartFill } from "react-icons/bs";
 import { FaChevronUp, FaChevronDown } from "react-icons/fa6";
 import { useNavigate } from "react-router-dom";
@@ -6,161 +5,82 @@ import styled from "styled-components";
 
 import Button from "@/components/Button";
 import SelectBox from "@/components/SelectBox";
+import { PRODUCT_DETAIL_INFO } from "@/constants";
+import useProductDetail from "@/hooks/useProductDetail";
 
 import OptionProductBox from "./OptionProduct";
 
-const productData: ProductDetail = {
-  id: 100,
-  name: "파타고니아 레트로 x 양털 후리스 뽀글이 플리스 자켓",
-  imageUrl: "string",
-  clothesInfo: [
-    {
-      id: 101,
-      color: "BROWN",
-      size: "S",
-    },
-    {
-      id: 102,
-      color: "BROWN",
-      size: "M",
-    },
-    {
-      id: 103,
-      color: "BROWN",
-      size: "L",
-    },
-    {
-      id: 104,
-      color: "BLACK",
-      size: "M",
-    },
-  ],
-  description: "상세정보1",
-  brandInfo: {
-    name: "파타고니아",
-    id: 1,
-  },
-  liked: false,
-  price: 198000,
-  discountPrice: 0,
-};
-
-interface ColorMap {
-  [key: string]: { id: number; size: string }[];
-}
-
-function createColorObject(productData: ProductDetail) {
-  const colorMap: ColorMap = {};
-
-  productData.clothesInfo.forEach((item: ClothesInfo) => {
-    const { id, color, size } = item;
-    if (!colorMap[color]) {
-      colorMap[color] = [];
-    }
-    colorMap[color].push({ id, size });
-  });
-  return colorMap;
-}
-
-const DetailInfoData = [
-  { id: 1, name: "상세정보", description: productData.description },
-  { id: 2, name: "배송안내", description: "배송안내2" },
-  { id: 3, name: "교환 및 반품안내", description: "교환 및 반품안내3" },
-  { id: 4, name: "품질 보증 및 A/S", description: "품질 보증 및 A/S4" },
-  { id: 5, name: "취급주의사항", description: "취급주의사항5" },
-];
-
 export default function DetailInfo() {
   const navigate = useNavigate();
-  const [selectedSizes, setSelectedSizes] = useState<string[]>([]);
-  const [toggleDetailInfo, setToggleDetailInfo] = useState([0]);
 
-  const [selectedColor, setSelectedColor] = useState<string | null>("");
-  const [selectedOptions, setSelectedOptions] = useState<ClothesInfo[]>([]);
-  const options = createColorObject(productData);
-  console.log(options);
-
-  useEffect(() => {
-    const selectedColorOption = selectedOptions.filter(
-      (option) => option.color === selectedColor,
-    );
-    setSelectedSizes(selectedColorOption.map((option) => option.size));
-  }, [selectedColor, selectedOptions]);
-
-  const handleSelection = (id: number, color: string, size: string) => {
-    if (!selectedColor) return alert("색상을 선택해주세요");
-    if (selectedSizes.includes(size)) {
-      setSelectedSizes(selectedSizes.filter((size) => size !== size));
-      setSelectedOptions(selectedOptions.filter((option) => option.id !== id));
-    } else {
-      setSelectedSizes([...selectedSizes, size]);
-      setSelectedOptions((prev) => [...prev, { id, color, size }]);
-    }
-  };
-
-  const handleRemoveOption = (optionId: number, optionSize: string) => {
-    setSelectedSizes(selectedSizes.filter((size) => size !== optionSize));
-    setSelectedOptions(
-      selectedOptions.filter((options) => options.id !== optionId),
-    );
-  };
-
-  const handleToggleDetailInfo = (infoId: number) => {
-    if (toggleDetailInfo.includes(infoId)) {
-      setToggleDetailInfo(toggleDetailInfo.filter((id) => id !== infoId));
-    } else {
-      setToggleDetailInfo([...toggleDetailInfo, infoId]);
-    }
-  };
+  const {
+    productData,
+    options,
+    selectedColor,
+    setSelectedColor,
+    handleSelectOption,
+    selectedSizes,
+    selectedOptions,
+    handleRemoveOption,
+    toggleDetailInfo,
+    handleToggleDetailInfo,
+  } = useProductDetail();
 
   return (
     <ProductDetail>
-      <ProductCode
-        onClick={() => navigate(`/brands/${productData.brandInfo.id}`)}
-        style={{ cursor: "pointer" }}
-      >
-        {productData.brandInfo.name}
-      </ProductCode>
-      <ProductInfo>{productData.name}</ProductInfo>
+      {productData && (
+        <>
+          <ProductCode
+            onClick={() => navigate(`/brands/${productData.brandInfo.id}`)}
+            style={{ cursor: "pointer" }}
+          >
+            {productData.brandInfo.name}
+          </ProductCode>
+          <ProductInfo>{productData.name}</ProductInfo>
 
-      <ProductPrice>₩ {productData.price.toLocaleString()}</ProductPrice>
-      <ProductColor>
-        <SelectBox
-          options={Object.keys(options)}
-          id="colors"
-          text="색상을 선택해주세요"
-          selected={selectedColor}
-          setSelected={setSelectedColor}
-          className="colorOption"
-        />
-      </ProductColor>
-      {selectedColor && (
-        <ProductSize>
-          {options[selectedColor]?.map((option) => (
-            <SizeBox
-              key={option.id}
-              onClick={() =>
-                handleSelection(option.id, selectedColor, option.size)
-              }
-              selected={selectedSizes.includes(option.size)}
-            >
-              {option.size}
-            </SizeBox>
-          ))}
-        </ProductSize>
+          <ProductPrice>₩ {productData.price.toLocaleString()}</ProductPrice>
+          <ProductColor>
+            {options && (
+              <SelectBox
+                options={Object.keys(options)}
+                id="colors"
+                text="색상을 선택해주세요"
+                selected={selectedColor}
+                setSelected={setSelectedColor}
+                className="colorOption"
+              />
+            )}
+          </ProductColor>
+          {selectedColor && (
+            <ProductSize>
+              {options &&
+                options[selectedColor]?.map((option) => (
+                  <SizeBox
+                    key={option.id}
+                    onClick={() =>
+                      handleSelectOption(option.id, selectedColor, option.size)
+                    }
+                    selected={selectedSizes.includes(option.size)}
+                  >
+                    {option.size}
+                  </SizeBox>
+                ))}
+            </ProductSize>
+          )}
+          {selectedOptions.length > 0 &&
+            selectedOptions.map((option) => (
+              <OptionProductBox
+                key={option.id}
+                price={productData.price}
+                selectedSize={option.size}
+                selectedColor={option.color}
+                handleRemoveOption={() =>
+                  handleRemoveOption(option.id, option.size)
+                }
+              />
+            ))}
+        </>
       )}
-      {selectedOptions.length > 0 &&
-        selectedOptions.map((option) => (
-          <OptionProductBox
-            key={option.id}
-            price={productData.price}
-            selectedSize={option.size}
-            selectedColor={option.color}
-            handleRemoveOption={() =>
-              handleRemoveOption(option.id, option.size)
-            }
-          />
-        ))}
       <ButtonBox>
         <Button $variant="solid" color="primary">
           구매하기
@@ -183,7 +103,7 @@ export default function DetailInfo() {
         </Button>
       </ButtonBox>
       <ProductDetailInfo>
-        {DetailInfoData.map((info) => (
+        {PRODUCT_DETAIL_INFO.map((info) => (
           <DetailInfoKey
             key={info.id}
             onClick={() => handleToggleDetailInfo(info.id)}
